# Performance MetaDataEditor When Selecting Many Images