Related Issues:
- #4196

In a meta data editor with 2000 images, the method [`DataEditorForm.consecutivePagesSelected`](https://github.com/kitodo/kitodo-production/blob/ed3b6c6301630e0d045156e49d3171b5328ae3ac/Kitodo/src/main/java/org/kitodo/production/forms/dataeditor/DataEditorForm.java#L689-L701) is called over 2000 times when selecting an image. For large gallery sections (>100 images) and large selections (>10 images), it's execution time is relatively slow (milliseconds), which adds up when called over 2000 times to multiple seconds.

For example, when selecting 50 images in a gallery section of 500 images:

https://user-images.githubusercontent.com/6214043/167098111-67aada5c-90f2-487e-ada8-27e367249a84.mp4

Or, when selecting:
- 1 image in a gallery section of 80 images: 21.6 ms. 
- 1 image in a gallery section of 500 images: 483ms
- 10 images in a gallery section of 80 images: 105ms
- 10 images in a gallery section of 500 images: 2.8 seconds
- 100 images in a gallery section of 500 images: 23.7 seconds

Most of it's execution time seems to be related to the [`distinct().count()`](https://github.com/kitodo/kitodo-production/blob/ed3b6c6301630e0d045156e49d3171b5328ae3ac/Kitodo/src/main/java/org/kitodo/production/forms/dataeditor/DataEditorForm.java#L696) call. For large selections, this should be slower, since it involves pairwise comparisons. However, I'm not sure why it is also slower for large gallery sections, since the size of a gallery section should not have any influence here.

Anyway, the proposed implementation improves the runtime of the method signficantly. The same selection of 50 images loads much faster:

https://user-images.githubusercontent.com/6214043/167098512-fc9994f6-9eaf-4504-a846-634fe494d4f5.mp4

The execution time of `DataEditorForm.consecutivePagesSelected` is reduced to about 36ms in the worst case (selecting 500 images in a gallery section of 500).

The remaining delay when selecting images is related to the full reloading of the gallery, structure and pagination panel.